Add quite a few changes, most of which improve the way data is collected
A "forwards" input, in which only W is being pressed, is only added to the data when there is another frame to go with it in which another key is being pressed. This means that "forwards" inputs make up about 50% of the data.

The model is trained to predict what the input will be four frames in the future. This approximately reflects the way a human drives: they press the key which is appropriate to what they saw about 200ms ago. When the model is running its prediction is stored for four frames before being used so that there is symmetry between the way it is trained and the way it has to drive.