A few improvements
I'm committing now in preparation for a change to the input and output formats.

Two main improvements:

1) The model is now called directly in collect_correction_data.py instead of the
.predict method being called. This resulted in a dramatic speed-up because .predict
is optimised for evaluating the model over large sets of data whereas calling it
directly is intended for small amounts.

2) The model is now specified with Keras' functional API rather than the sequential
API. The functional API can be used to specify models with multiple inputs and
outputs.

I made a model with two inputs: the current frame and the keys being pressed on the
last 28 frames. This resulted in a model which almost never changed the key which
it was pressing. If I corrected it, it would just carry on with the key I had
pressed. After looking at the data briefly I concluded that this is because on
about 85% of frames the key currently being pressed is the same as the key being
pressed on the previous frame.