Restructure and neaten the repo
Deleted lots of unused files.

Each model format will now have its own directory in which the code appears in
a set of files, each devoted to a particular part of the workflow (as opposed
one big module containing all of the code).
key_press_frames_to_multihot_4 is an example of this. The only file that I've
verified to still be working is run_model.py. frame_seq_to_mh4 will also be
re-structured in this way if and when I go back to that model format.