Add final tweaks to TF-Agents code
TF-Agents has not proven to be useful so I'm moving on. I think that the agent is gradually learning over the course of several million iterations, but learning stops after restoring from a checkpoint. This problem, along with the fact that the agent doesn't seem to reduce epsilon, make me feel like it will not be worthwhile continuing with TF-Agents.