Change the default values for the numbers environment
Also clean up the file a bit, and remove numbers_pyenv which was only used with TFAgents.

The new default values for numbers_env make the discounted reward graphs look more familiar. With
a delay of 0 and sparsity of 1 the graphs looked incorrect, which may be a bug; I have not
investigated.